+++ app/controllers/admin/territories/zones_controller.rb
@@ -20,12 +20,12 @@ class Admin::Territories::ZonesController < Admin::Territories::BaseController
     zone_defaults = { level: params[:default_zone_level] || Zone::LEVEL_CITY }
     @zone = Zone.new(**zone_defaults.merge(zone_params_get), sector: @sector)
     @sectors = sector_policy.resolve
-    authorize_agent @zone
+    authorize(@zone, policy_class: Agent::ZonePolicy)
   end
 
   def create
     @zone = Zone.new(**zone_params, sector: @sector)
-    authorize_agent @zone
+    authorize(@zone, policy_class: Agent::ZonePolicy)
     if @zone.save
       if params[:commit] == I18n.t("helpers.submit.create")
         redirect_to admin_territory_sector_path(current_territory, @sector), flash: { success: "#{@zone.human_attribute_value(:level)} ajoutée au secteur" }
@@ -39,7 +39,7 @@ class Admin::Territories::ZonesController < Admin::Territories::BaseController
 
   def destroy
     zone = Zone.find(params[:id])
-    authorize_agent zone
+    authorize(zone, policy_class: Agent::ZonePolicy)
     if zone.destroy
       redirect_to admin_territory_sector_path(current_territory, @sector), flash: { success: "#{zone.human_attribute_value(:level)} retirée du secteur" }
     else
@@ -49,7 +49,7 @@ class Admin::Territories::ZonesController < Admin::Territories::BaseController
 
   def destroy_multiple
     zones = @sector.zones
-    zones = zones.filter { |z| authorize_agent(z, :destroy?) }
+    zones = zones.filter { |z| authorize(z, :destroy?, policy_class: Agent::ZonePolicy) }
     count = zones.count
     if zones.map(&:destroy).all?
       flash[:success] = "Les #{count} communes et rues ont été retirées du secteur"
